Split FenceTracker tests into module
